Typo
Me falto convertir los strlen en int

+++ batallas.c
@@ -41,8 +41,8 @@ int funcion_batalla_4(void* pkm_1, void* pkm_2){
 int funcion_batalla_5(void* pkm_1, void* pkm_2){
 
     // la velocidad se ve disminuida por lo que tardan en decir su nombre
-    int modificador_1 = ((pokemon_t*)pkm_1)->velocidad / strlen(((pokemon_t*)pkm_1)->nombre);
-    int modificador_2 = ((pokemon_t*)pkm_2)->velocidad / strlen(((pokemon_t*)pkm_2)->nombre);
+    int modificador_1 = ((pokemon_t*)pkm_1)->velocidad / (int) strlen(((pokemon_t*)pkm_1)->nombre);
+    int modificador_2 = ((pokemon_t*)pkm_2)->velocidad / (int) strlen(((pokemon_t*)pkm_2)->nombre);
 
     int ataque_1 = ((pokemon_t*)pkm_1)->ataque * modificador_1;
     int ataque_2 = ((pokemon_t*)pkm_2)->ataque * modificador_2;
